🎨: デザインシステム連携のため余白応用パターンの記述構成を見直す
